Reworked gap as int, log more info to .txt information

+++ src/main/kotlin/org/jetbrains/bio/span/SpanCLAAnalyze.kt
@@ -110,7 +110,7 @@ object SpanCLAAnalyze {
                 require(sensitivity == null || 0 < sensitivity && sensitivity <= 1) {
                     "Illegal background sensitivity: $sensitivity, expected range: (0, 1]"
                 }
-                val gap = if (options.has("gap")) options.valueOf("gap") as Double else null
+                val gap = if (options.has("gap")) options.valueOf("gap") as Int else null
                 require(gap == null || gap >= 0) { "Illegal gap: $gap, expected >= 0" }
 
                 val workingDir = options.valueOf("workdir") as Path
