+++ src/main/kotlin/org/jetbrains/bio/span/peaks/SpanPeaksResult.kt
@@ -5,7 +5,7 @@ import org.jetbrains.bio.genome.containers.GenomeMap
 data class SpanPeaksResult(
     val fdr: Double,
     val sensitivity: Double,
-    val gap: Double,
+    val gap: Int,
     val peaks: GenomeMap<List<Peak>>
 ) {
     fun toList(): List<Peak> = peaks.genomeQuery.get().flatMap { peaks[it] }
